major refactor completed, created folders for images,json,components
