+++ src/components/markdownComponent.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import ReactMarkdown from 'react-markdown';
+
+const MarkdownComponent = ({ markdownContent }) => {
+  return (
+    <div>
+      <ReactMarkdown>{markdownContent}</ReactMarkdown>
+    </div>
+  );
+};
+
+export default MarkdownComponent;
